pyproject.toml
[tool.poetry]
name = "reflexio"
version = "0.1.2"
description = "A Python library for the Reflexio"
authors = ["Reflexio Team"]
readme = "README.md"
packages = [{include = "reflexio"}]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.1"
psycopg2-binary = "^2.9.9"
boto3 = "^1.34.149"
sqlalchemy = "^2.0.31"
python-jose = "^3.3.0"
passlib = "^1.7.4"
openai = ">=2.8.0"
tenacity = "^9.0.0"
duckduckgo-search = "^7.0.1"
bcrypt = "^4.2.1"
reflexio-commons = {path = "reflexio/reflexio_commons", develop = true}
uvicorn = "^0.34.0"
xlsxwriter = "^3.2.2"
aiohttp = "^3.11.12"
supabase = "^2.15.1"
python-dotenv = "^1.1.0"
hdbscan = "^0.8.40"
redis = "^6.2.0"
websocket-client = "^1.8.0"
tiktoken = "^0.12.0"
anthropic = "^0.72.0"
litellm = "^1.80.11"
slowapi = "^0.1.9"
cachetools = "^6.2.4"
colorlog = "^6.10.1"
httpx = "^0.28.1"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.5.3"
mkdocs-material = "^9.5.3"
mkdocstrings = {extras = ["python"], version = "^0.24.0"}
griffe = "^0.38.0"
mkdocstrings-python = "^1.7.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.0"
pytest-xdist = "^3.8.0"
black = "^24.10.0"
moto = "^5.0.28"
jupyterlab = "^4.4.3"
matplotlib = "^3.10.8"
ruff = ">=0.15.0"
pyright = ">=1.1.400"

[tool.pytest.ini_options]
# Parallel test execution with pytest-xdist
# -n auto: Use all available CPU cores
# --dist loadfile: Group tests by file (reduces fixture overhead, safer for in-file dependencies)
addopts = "-n auto --dist loadfile"
testpaths = ["reflexio/tests"]
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes (unused imports, undefined names)
    "W",      # pycodestyle warnings
    "B",      # flake8-bugbear (likely bugs & design problems)
    "UP",     # pyupgrade (modern Python syntax)
    "I",      # isort (import sorting)
    "C4",     # flake8-comprehensions (correct dict/list/set usage)
    "SIM",    # flake8-simplify (code simplification patterns)
    "RET",    # flake8-return (return statement quality)
    "PERF",   # perflint (performance anti-patterns)
    "C90",    # McCabe complexity
    "S",      # flake8-bandit (security checks)
    "N",      # pep8-naming conventions
    "ANN",    # flake8-annotations (enforce type annotations)
    "ARG",    # flake8-unused-arguments
    "G",      # flake8-logging-format (no f-strings in logging)
    "PTH",    # flake8-use-pathlib (pathlib over os.path)
    "FURB",   # refurb (modern Python idioms)
    "FLY",    # flynt (string join → f-string)
    "PLR1714",# pylint: repeated-equality-comparison → use `in`
    "PLR1716",# pylint: boolean-chained-comparison → chain
]
ignore = [
    "E501",   # line too long (handled by formatter/black)
    "E402",   # module-level import not at top (common in __init__.py)
    "ANN002",  # missing type for *args (often inherited from frameworks)
    "ANN003",  # missing type for **kwargs (often inherited from frameworks)
    "ANN204",  # missing return type for __init__ (-> None is obvious)
    "ANN401",  # dynamically typed Any (sometimes intentional)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]
"reflexio/tests/**" = ["S101", "ARG001", "ARG002", "PTH", "ANN"]
"conftest.py" = ["ARG001"]
"reflexio/reflexio_commons/tests/**" = ["S101", "ARG001", "ARG002", "PTH", "ANN"]
"notebooks/**" = ["S101", "F821", "S105", "S106", "ANN"]
"reflexio/scripts/**" = ["S101", "G004", "ANN"]
"reflexio/reflexio_client/tests/**" = ["S101", "ARG001", "ARG002", "PTH", "ANN"]
"reflexio/server/api.py" = ["B008", "ARG001", "ARG002"]
"reflexio/server/api_endpoints/**" = ["B008", "ARG001", "ARG002"]
"reflexio/server/services/configurator/test_config_storage.py" = ["S101", "ANN"]
"demo/**" = ["S101", "PTH", "G004", "ANN"]

[tool.ruff.lint.mccabe]
max-complexity = 20

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
